Adding authentication check for delete questions

+++ src/ConsultBundle/Controller/QuestionsController.php
@@ -5,6 +5,7 @@ namespace ConsultBundle\Controller;
 use FOS\RestBundle\View\View;
 use FOS\RestBundle\Util\Codes;
 use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use ConsultBundle\Manager\ValidationError;
 
@@ -132,6 +133,12 @@ class QuestionsController extends BaseConsultController
 
     public function deleteQuestionAction($id)
     {
+        $practoAccountId = $this->authenticate();
+
+        if ($practoAccountId != ) {
+           throw new HttpException(Codes::HTTP_FORBIDDEN, "Unauthorised Access");
+        }
+
         $questionManager = $this->get('consult.question_manager');
         $questionManager->delete($id);
     }
